requirements.txt
aiogram
google-generativeai
pydub
speech_recognition
fastapi
uvicorn
